error handling on abckend and fix refresh rate on frontend
