+++ frontend/src/api/k8s/metricsData.ts
@@ -2,6 +2,7 @@ import axios from 'axios';
 import { PrometheusQueryResponse } from '~/types';
 import usePrometheusQuery from '../prometheus/usePrometheusQuery';
 import React from 'react';
+import { timeStringToSeconds } from '~/pages/MCADashboard/Metrics/metrics-utils';
 
 const getMetricData = async (query: string) => {
   const body = { query: query };
@@ -20,26 +21,6 @@ const getMetricDataRange = async (query: string, range: string) => {
   body.range = rangeArr;
   const res = await axios.post('/api/metrics-data', body);
   return res;
-};
-
-const timeStringToSeconds = (timeString: string) => {
-  const value: number = parseInt(timeString.substring(0, timeString.length - 1));
-  const unit = timeString.charAt(timeString.length - 1).toLowerCase();
-  console.log('timeString', timeString);
-  switch (unit) {
-    case 'm':
-      return value * 60;
-    case 'h':
-      return value * 60 * 60;
-    case 'd':
-      return value * 24 * 60 * 60;
-    case 'w':
-      return value * 7 * 24 * 60 * 60;
-    case 'M':
-      return value * 7 * 24 * 60 * 60 * 4;
-    default:
-      throw new Error('Invalid time unit');
-  }
 };
 
 export { getMetricData, getMetricDataRange };
